requirements.txt
Flask==2.3.3
python-dotenv==1.0.1
openai>=1.0.0
gunicorn>=20.0.0,<23.0.0